% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMoonSimParameters.R
\name{buildMoonSimJobsList}
\alias{buildMoonSimJobsList}
\title{Function to return input parameters for main simulation.}
\usage{
buildMoonSimJobsList()
}
\description{
Function to return input parameters for main simulation.
}
