% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stealTheMoonSim.R
\name{stealTheMoonSim}
\alias{stealTheMoonSim}
\title{Simple example simulation for use in demonstrating rminions package.}
\usage{
stealTheMoonSim(fuelParams, solarFlareParams, flareSurfaceParams,
  flareAdditionalFuelParams, surfaceFuelDependenceCopParams, asteroidParams,
  asteroidAdditionalFuelParams)
}
\description{
Simple example simulation for use in demonstrating rminions package.
}
